BLE: Move Gap event types in dedicated file.
